+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- **httperr:** Moved HTTP related code to separate package
+
 
 ## [0.9.1] - 2018-07-27