+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **bugsnag:** Improve logger
 - **errorlog:** Renamed log package
 - **errorlog:** Improved package
+- **errorlogrus:** Rename logrus package
 
 
 ## [0.9.1] - 2018-07-27