+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.10.0] - 2018-08-21
+
 ### Added
 
 - **errorlogrus:** Add `AttachContext` option to the Hook so that the entry data is appended to the error
@@ -190,7 +193,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [Bugsnag](https://bugsnag.com) handler
 
 
-[Unreleased]: https://github.com/goph/emperror/compare/v0.9.1...HEAD
+[Unreleased]: https://github.com/goph/emperror/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/goph/emperror/compare/v0.9.1...v0.10.0
 [0.9.1]: https://github.com/goph/emperror/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/goph/emperror/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/goph/emperror/compare/v0.7.1...v0.8.0